Basic (almost) implementation of elastic GMLS functional.
* Renamed elasticity_gmls_functional -> elastic_gmls_functional.
* Implemented calculation of D, N, P matrices.
* Computation of functional from matrices.
